extract promise logic to an external helper
